son_blog/requirements.txt
appdirs==1.4.3
asgiref==3.2.4
distlib==0.3.0
Django==3.0.4
django-ckeditor==5.9.0
django-js-asset==1.2.2
filelock==3.0.12
Pillow==7.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.9
